% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_features.R
\name{check_column_class}
\alias{check_column_class}
\title{Check input data column classes}
\usage{
check_column_class(.x, .map)
}
\arguments{
\item{.x}{a \code{data.frame} to be checked}

\item{.map}{a two-column \code{data.frame} mapping column names to
(character vectors specifying) expected column classes.
The first column must be named \code{colname} and have type character.
The second column must be a list-column of character vectors and named \code{accepted_classes}.
(see \code{?\link{required.tweets.df.cols}}).}
}
\value{
a named logical vector with length equal to \code{length(.x)}.
    Each element is corresponds to a column of \code{.x} (names correspond).
    An element is \code{TRUE}, if the class of the column in \code{.x} matches
     any of the classes specified in \code{.map[.map$colname == <col>, "accepted_classes"]}.
    It is \code{NA}, if a column of \code{.x} is not mapped in \code{.map}
    Otherwise, it is \code{FALSE}.
}
\description{
Internal helper to \code{\link{create_tweet_features}}:
     Function takes a data frame \code{.x}
     and another data frame \code{.map} that
     maps expected column names to vectors specifying columns' accepted/expected
     classes, and checks that all columns in \code{.x}
     have the expected classes specified in  \code{.map}
}
