% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_tweets.R
\name{classify_tweets_debug_warning}
\alias{classify_tweets_debug_warning}
\title{Raise \code{\link{classify_tweets}} debugging warning}
\usage{
classify_tweets_debug_warning(.x, model)
}
\arguments{
\item{.x}{input data frame}

\item{.ensemble}{\code{\link[caretEnsemble]{caretEnsemble}} object}
}
\value{
none (raises warning invisibly)
}
\description{
Internal helper to \code{\link{classify_tweets}}
}
