% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_tweets.R
\name{classify_tweets_predict}
\alias{classify_tweets_predict}
\title{\code{classify_tweets} core module}
\usage{
classify_tweets_predict(
  model,
  x,
  threshold,
  .predict.type,
  .add,
  .debug,
  .verbose,
  ...
)
}
\arguments{
\item{model}{an 'caretEnsemble' object}

\item{x}{a 'data.frame' object (see \code{?\link{classify_tweets}})}

\item{threshold}{double in (0,1) (see \code{?\link{classify_tweets}})}

\item{.predict.type}{"prob" or "raw" (see \code{?\link{classify_tweets}})}

\item{.add}{logical (see \code{?\link{classify_tweets}})}

\item{.debug}{logical (see \code{?\link{classify_tweets}})}

\item{.verbose}{logical (see \code{?\link{classify_tweets}})}

\item{...}{arguments passed to predict method (see \code{?\link{classify_tweets}})}
}
\value{
a data.frame
}
\description{
Does part of the heavy lifting for \code{classify_tweets}'s methods
}
