% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tweet_features.R
\name{compute_tweet_features}
\alias{compute_tweet_features}
\title{Compute tweet features}
\usage{
compute_tweet_features(
  x,
  .as.data.table = TRUE,
  .mentions.regexp = "(?<=^|\\\\W)(@\\\\w{1,15})(?=\\\\s|$|\\\\W)",
  .hashtags.regexp = "(?<=\\\\.|^|\\\\s)(#\\\\w{1,139})(?=\\\\s|$|\\\\W)",
 
    .url.regexp = "\\\\b(([a-z][\\\\w-]+:(/{1,3}|[a-z0-9\%])|www\\\\d{0,3}[.]|[a-z0-9.\\\\-]+[.][a-z]{2,4}/)([^\\\\s()<>]+|\\\\(([^\\\\s()<>]+|(\\\\([^\\\\s()<>]+\\\\)))*\\\\))+(\\\\(([^\\\\s()<>]+|(\\\\([^\\\\s()<>]+\\\\)))*\\\\)|[^\\\\s`!()\\\\[\\\\]{};:\\\\'\\\\\\".,<>?«»“”‘’]))"
)
}
\arguments{
\item{x}{a \code{data.frame} \code{\link[data.table]{data.table}}, or \code{\link[tibble]{tibble}}
recording tweets.
For required column (naming and typing conventions) refer to \code{?\link{required.tweets.df.cols}}.
For an example see \code{?\link{tweets.df.prototype}}.}

\item{.as.data.table}{logical. Whether or not to return a \code{\link[data.table]{data.table}}.
Defaults to \code{TRUE}.
If \code{FALSE}, the returned object is a \code{\link[tibble]{tibble}}.}

\item{.mentions.regexp}{unit-length character vector,
specifying the regular expression pattern used to match, count, and remove
mentions in the tweet text (applied to column \code{text}).}

\item{.hashtags.regexp}{unit-length character vector,
specifying the regular expression pattern used to match, count, and remove
hashtags in the tweet text (applied to column \code{text}).}

\item{.url.regexp}{unit-length character vector,
specifying the regular expression pattern used to match, count, and remove
URLs in the tweet text (applied to column \code{text}).}
}
\value{
A \code{\link[data.table]{data.table}} if \code{.as.data.table = TRUE} (default),
     otherwise a \code{\link[tibble]{tibble}}.
     The return object contains all columns contained in \code{x} \emph{plus}
      the created tweet features.
}
\description{
Internal helper to \code{\link{create_tweet_features}}
}
