% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constituent.models}
\alias{constituent.models}
\title{Constituent model (pre-trained base learners)}
\format{
A list of \code{\link[caret]{train}} objects.
   \itemize{
     \item{\code{glmnet}: a generalized linear model (GLM) with Elastic-Net regularization (\code{\link[glmnet]{glmnet}})}
     \item{\code{svmRadial}: a Support Vector Machine (SVM) with a radial kernel (\code{\link[kernlab]{ksvm}} with \code{kernel = "rbfdot"})}
     \item{\code{ranger}: a Random Forest (\code{\link[range]{ranger}})}
     \item{\code{xgbTree}: an eXtreme Gradient Boosting (XGBoost) machine (\code{\link[xgboost]{xgboost}} with \code{learner = "tree"})}
   }
}
\usage{
constituent.models
}
\description{
List of constituent models trained with \code{\link[caretEnsemble]{caretList}}.
}
\details{
Individual models were tuned to individually defined tuning grids using
      5-times repeated 10-fold cross validation to 6292 labeled tweets.
     "Best" models were selected among these candidates based on which
      tuning parameter combinations yielded the highest F1 value.
     These "best" candidates were then trained to the complete training data
      in a 10-times repeated 10-fold cross validation scheme.
}
\keyword{datasets}
