% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_missings}
\alias{detect_missings}
\title{Detect missings}
\usage{
detect_missings(.x)
}
\arguments{
\item{.x}{a \code{data.frame}}
}
\value{
A list with elements
    \describe{
      \item{removed}{
        An integer vector of indexes of rows with any \code{NA}, \code{NaN} or (±) \code{Inf}
        \code{NULL} if all rows are complete (all values non-missings).
       }
      \item{rows_na_map}{
        \code{NULL}  if \code{removed} is \code{NULL}.
        Otherwise, a named list (names correspond to \code{removed}).
        Each list element is a list with elements "na", "nan" and "inf",
        each a character vector of column names with
        \code{NA}, \code{NaN} and (±) \code{Inf} values, respectively.
      }
    }
}
\description{
Function detects rows of for which any column is \code{NA}, \code{NaN} or (±) \code{Inf}.
}
