% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensemble.model}
\alias{ensemble.model}
\title{Ensemble model (pre-trained ensemble classifier)}
\format{
A 'caretEnsemble' object.
   Element 'models' is identical to \code{\link{constituent.models}}
   Element 'ens_model' is the ensemble classifier induced by PR-AUC-based blending.
}
\usage{
ensemble.model
}
\description{
Pre-trained ensemble classifier blended with \code{metric = "PR-AUC"}
    \code{\link{constituent.models}} using \code{\link[caretEnsemble]{caretEnsemble}}.
}
\details{
Individual models were tuned to individually defined tuning grids using
      5-times repeated 10-fold cross validation to 6292 labeled tweets.
     "Best" models were selected among these candidates based on which
      tuning parameter combinations yielded the highest F1 value.
     These "best" candidates were then trained to the complete training data
      in a 10-times repeated 10-fold cross validation scheme.

     The ensemble classifier was then obtained using \code{\link[caretEnsemble]{caretEnsemble}}
      to blend them according to their test set performance (measured by PR-AUC) in these 100 resamples.
}
\keyword{datasets}
