% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laser.embedding.icomp}
\alias{laser.embedding.icomp}
\title{300 independent components of LASER representations of tweets in labeled dataset}
\format{
A \code{\link[fastICA]{fastICA}} object, which is a list with the following elements:
    \describe{
       \item{K}{The 1024x300 'pre-whitening' \code{matrix} projecting embeddings onto the first 300 principal components}
       \item{W}{The 300x300 'un-mixing' \code{matrix} estimated from the original model input data}
       \item{A}{The 300x1024 'mixing' \code{matrix} estimated from the original model input data}
       \item{X.means}{a \code{double} vector with 1024 elements recording embedding vector means in the original model input data}
       \item{runtime}{\code{difftime} object reporting the time (in seconds) whitening and un-mixing took}
       \item{params}{a list mapping the parameter used when fitting \code{\link[fastICA]{fastICA}}}
     }
}
\usage{
laser.embedding.icomp
}
\description{
\code{\link[fastICA]{fastICA}} object fitted to the LASER embedding representations
    of tweet texts contained in the training and validation data used
    to obtain the ensemble classifier distributed with
    the \code{politicaltweets} package
}
\examples{
\dontrun{
 # create fake 'test' observation (1-row matrix)
 X <- matrix(
   rnorm(1024)
   , nrow = 1
   # column names must be "e0001", ..., "e1024"
   , dimnames = list(c("test"), sprintf("e\%04d", 1:1024))
 )
 #'  # obtain independent component representation of x
 X \%*\% laser.embedding.icomp$K \%*\% laser.embedding.icomp$W
}
}
\keyword{datasets}
