% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laser.embedding.prcomp}
\alias{laser.embedding.prcomp}
\title{First 300 principal components of LASER representations of tweets in labeled dataset}
\format{
A \code{\link[stats]{prcomp}} object, which is a list with the following elements:
    \describe{
       \item{sdev}{standard deviations of the principal components}
       \item{rotation}{a \code{matrix} of variable loadings. Row names "e0001", ..., "e1024"; column names "PC1", ..., "PC300"  }
       \item{center}{\code{FALSE}, since no centering applied}
       \item{scale}{\code{FALSE}, since no scaling applied}
     }
}
\usage{
laser.embedding.prcomp
}
\description{
\code{\link[stats]{prcomp}} object fitted to the LASER embedding representations
    of tweet texts contained in the training and validation data used
    to obtain the ensemble classifier distributed with
    the \code{politicaltweets} package
}
\examples{
\dontrun{
 # create fake 'test' observation (1-row matrix)
 X <- matrix(
   rnorm(1024)
   , nrow = 1
   # column names must be "e0001", ..., "e1024"
   , dimnames = list(c("test"), sprintf("e\%04d", 1:1024))
 )

 # use precomp predict method to obtain PCs
 predict(laser.embedding.prcomp, newdata = ndat)
}
}
\keyword{datasets}
