% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{required.tweets.df.cols}
\alias{required.tweets.df.cols}
\title{Required tweets data frame columns}
\format{
A \code{\link[tibble]{tibble}} with columns
     \describe{
       \item{colname}{a character vector, specifying required column names}
       \item{accepted_types}{a list column of character vectors, each element mapping the accepted classed to \code{colname}}
     }
}
\usage{
required.tweets.df.cols
}
\description{
Required tweets data frame columns
}
\keyword{datasets}
