% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect_quote.R
\name{str_detect_one_quote_pattern}
\alias{str_detect_one_quote_pattern}
\title{Detect a quote pattern in text}
\usage{
str_detect_one_quote_pattern(x, pattern, .verbose = FALSE)
}
\arguments{
\item{x}{character vector}

\item{pattern}{a list with elements
"s" (unit-length character specifying regex to match quote start)
, "e" (unit-length character specifying regex to match quote end)
, and "min_words" unit-length integer specifying the min. No. words to count a match as a quote)
See \code{\link{text.quote.patterns}}.}

\item{.verbose}{logical. verbosity}
}
\value{
A list with elements
   \describe{
       \item{quotes}{logical indicating whether or not \code{x} contains a quote that matches the quote pattern (\code{pattern})}
       \item{n_quotes}{non-negative integer, the number of quotes detected }
       \item{n_words}{non-negative integer, the number of words contained in the quote}
       \item{n_chars}{non-negative integer, the number of characters contained in the quote}
   }
}
\description{
Detect a quote pattern in text
}
