% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect_quote.R
\name{str_detect_quote}
\alias{str_detect_quote}
\title{Detect quotes in strings}
\usage{
str_detect_quote(x, quote.patterns = text.quote.patterns, verbose = FALSE)
}
\arguments{
\item{x}{character vector of strings to be parsed}

\item{quote.patterns}{list of list of quote patterns to be detected (see \code{\link{text.quote.patterns}})}

\item{verbose}{logical}
}
\value{
a list object
}
\description{
Detect quotes in strings
}
