% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect_quote.R
\name{str_detect_quote_one}
\alias{str_detect_quote_one}
\title{Detect quotes in a string}
\usage{
str_detect_quote_one(
  x,
  .quote.patterns = text.quote.patterns,
  .verbose = FALSE
)
}
\arguments{
\item{x}{unit-length character vector to be parsed}

\item{.quote.patterns}{list of list of qquote patterns to be detected (see \code{\link{text.quote.patterns}})}

\item{.verbose}{logical}
}
\value{
a list object
}
\description{
Detect quotes in a string
}
