% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_statistics.R
\name{str_emoji_stats}
\alias{str_emoji_stats}
\title{Compute Emoji statistics for string}
\usage{
str_emoji_stats(x, emojis.vec = rtweet::emojis$code)
}
\arguments{
\item{x}{character vector, vectorized over elements}

\item{emojis.vec}{character vector specifying Emoji Unicode values to
be matched and extracted.
Defaults to column "code" of the \code{\link[rtweet]{emojis}} data.}
}
\value{
A list (of lists) with elements
   \describe{
       \item{ratio}{the ratio of Emoji characters relative to all characters of the string}
       \item{count}{the number of Emoji characters}
       \item{emojis}{a character vector recording all Emojis matched in (the current element of) `x`}
   }
}
\description{
Compute Emoji statistics for string
}
