% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_statistics.R
\name{str_pattern_width}
\alias{str_pattern_width}
\title{Compute number of characters matching pattern}
\usage{
str_pattern_width(x, pattern)
}
\arguments{
\item{x}{character vector, vectorized over elements}

\item{pattern}{unit-length character vector, a pattern (regex) to be matched in \code{x}}
}
\value{
An numeric (integer) vector
}
\description{
Compute number of characters matching pattern
}
