% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_statistics.R
\name{str_unicode_char_ratio}
\alias{str_unicode_char_ratio}
\title{Compute Unicode character ratio}
\usage{
str_unicode_char_ratio(x)
}
\arguments{
\item{x}{character vector, vectorized over elements}
}
\value{
A numeric (double) vector
}
\description{
Compute Unicode character ratio
}
