% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_statistics.R
\name{str_unicode_count}
\alias{str_unicode_count}
\title{Compute Unicode character count}
\usage{
str_unicode_count(x)
}
\arguments{
\item{x}{character vector, vectorized over elements}
}
\value{
A numeric (integer) vector
}
\description{
Compute Unicode character count
}
