% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_column_mapping}
\alias{validate_column_mapping}
\title{Validate column mapping}
\usage{
validate_column_mapping(col.map)
}
\arguments{
\item{col.map}{expects a valid "column mapping",
that is a data frame that
\enumerate{
   \item{inherits from \code{data.frame}}
   \item{has two columns named "colname" and "accepted_classes"}
   \item{"colname" is type character and "accepted_classes" is a list column}
   \item{each element of "accepted_classes" is a character vector}
   \item{the \code{lengths} of "accepted_classes" is ≥ 1 for all its elements}
}}
}
\value{
a list with two elements
    \describe{
       \item{result}{a logical indicating whether \code{col.map} is a valid column mapping data frame}
       \item{message}{a character vector, recording the reason why \code{col.map} is invalid if \code{result} is \code{FALSE}}
    }
}
\description{
Function validates a column mapping
}
