% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_tweets.R
\name{validate_train_ctrl_sum_fun}
\alias{validate_train_ctrl_sum_fun}
\title{Validate \code{trainControl} summary function}
\usage{
validate_train_ctrl_sum_fun(fun, levs, metric)
}
\arguments{
\item{fun}{summary function, needs to conform to \code{caret} standards see \code{?\link[caret]{twoClassSummary}}}

\item{levs}{character with two elements, specifying outcome classes}

\item{metric}{evaluation metric expected in \code{fun}'s return object}
}
\value{
a list with elements "result" (unit-length logical) and "message" (character)
}
\description{
Internal helper to \code{\link{classify_tweets}}
}
\examples{
\dontrun{
library(caret)
# returns "result" FALSE and a "message"
validate_train_ctrl_sum_fun(twoClassSummary, c("yes", "no"), "A")
# returns "result" TRUE and no "message"
validate_train_ctrl_sum_fun(twoClassSummary, c("yes", "no"), "ROC")
}
}
